Make OnboardAPITests buildable.
